[rpm] Package minidlna for Sailfish use
Basic packaging for minidlna to be used in Jolla phone / sailfish.

Currently there is some issue with the server not showing at least
in Ubuntu's UPnP Inspector tool. Some more testing needed to see
if this can actually work.

Runs now as root service, but could be made as a user service with
some effort.

Signed-off-by: Kalle Jokiniemi <kalle.jokiniemi@jolla.com>